+++ presentation/diarywrite/src/main/java/com/hilingual/presentation/diarywrite/DiaryFeedbackLoadingScreen.kt
@@ -1,6 +1,5 @@
 package com.hilingual.presentation.diarywrite
 
-import androidx.annotation.RawRes
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
@@ -27,26 +26,7 @@ import com.hilingual.core.designsystem.component.button.HilingualButton
 import com.hilingual.core.designsystem.component.image.HilingualLottieAnimation
 import com.hilingual.core.designsystem.component.topappbar.CloseOnlyTopAppBar
 import com.hilingual.core.designsystem.theme.HilingualTheme
-
-internal enum class DiaryWriteFeedbackState(
-    val title: String,
-    val description: String,
-    @RawRes val lottieRawRes: Int,
-    val lottieRawResHeightDp: Dp
-) {
-    Loading(
-        title = "일기 저장 중..",
-        description = "피드백을 요청하고 있어요.",
-        lottieRawRes = R.raw.lottie_feedback_loading,
-        lottieRawResHeightDp = 194.dp
-    ),
-    Complete(
-        title = "AI 피드백 완료!",
-        description = "틀린 부분을 고치고,\n더 나은 표현으로 수정했어요!",
-        lottieRawRes = R.raw.lottie_feedback_complete,
-        lottieRawResHeightDp = 180.dp
-    )
-}
+import com.hilingual.presentation.diarywrite.component.DiaryWriteFeedbackState
 
 @Composable
 private fun DiaryFeedbackLoadingScreen(
